requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-Mail==0.10.0
python-dotenv==0.19.2
celery==5.2.3
requests==2.26.0
kombu==5.2.3
amqp==5.0.9

